refactor _get_valid_dtype to class method
