+++ woodwork/tests/logical_types/test_logical_types.py
@@ -1,6 +1,17 @@
 import pytest
 
-from woodwork.logical_types import Boolean, Categorical, Datetime, Ordinal
+from woodwork.accessor_utils import init_series
+from woodwork.logical_types import (
+    Boolean,
+    Categorical,
+    Datetime,
+    NaturalLanguage,
+    Ordinal
+)
+from woodwork.utils import import_or_none
+
+dd = import_or_none('dask.dataframe')
+ks = import_or_none('databricks.koalas')
 
 
 def test_logical_eq():
@@ -40,3 +51,78 @@ def test_ordinal_init_with_order():
     order = ('bronze', 'silver', 'gold')
     ordinal_from_tuple = Ordinal(order=order)
     assert ordinal_from_tuple.order == order
+
+
+def test_get_valid_dtype(sample_series):
+    valid_dtype = Categorical._get_valid_dtype(type(sample_series))
+    if ks and isinstance(sample_series, ks.Series):
+        assert valid_dtype == 'string'
+    else:
+        assert valid_dtype == 'category'
+
+    valid_dtype = Boolean._get_valid_dtype(type(sample_series))
+    assert valid_dtype == 'bool'
+
+
+def test_init_series_valid_conversion_specified_ltype(sample_series):
+    if ks and isinstance(sample_series, ks.Series):
+        sample_series = sample_series.astype('str')
+    else:
+        sample_series = sample_series.astype('object')
+
+    series = init_series(sample_series, logical_type='categorical')
+    assert series is not sample_series
+    correct_dtype = Categorical._get_valid_dtype(type(sample_series))
+    assert series.dtype == correct_dtype
+    assert isinstance(series.ww.logical_type, Categorical)
+    assert series.ww.semantic_tags == {'category'}
+
+    series = init_series(sample_series, logical_type='natural_language')
+    assert series is not sample_series
+    correct_dtype = NaturalLanguage._get_valid_dtype(type(sample_series))
+    assert series.dtype == correct_dtype
+    assert isinstance(series.ww.logical_type, NaturalLanguage)
+    assert series.ww.semantic_tags == set()
+
+
+def test_init_series_valid_conversion_inferred_ltype(sample_series):
+    if ks and isinstance(sample_series, ks.Series):
+        sample_series = sample_series.astype('str')
+    else:
+        sample_series = sample_series.astype('object')
+
+    series = init_series(sample_series)
+    assert series is not sample_series
+    correct_dtype = Categorical._get_valid_dtype(type(sample_series))
+    assert series.dtype == correct_dtype
+    assert isinstance(series.ww.logical_type, Categorical)
+    assert series.ww.semantic_tags == {'category'}
+
+
+def test_init_series_with_datetime(sample_datetime_series):
+    series = init_series(sample_datetime_series, logical_type='datetime')
+    assert series.dtype == 'datetime64[ns]'
+    assert isinstance(series.ww.logical_type, Datetime)
+
+
+def test_init_series_all_parameters(sample_series):
+    if ks and isinstance(sample_series, ks.Series):
+        sample_series = sample_series.astype('str')
+    else:
+        sample_series = sample_series.astype('object')
+
+    metadata = {'meta_key': 'meta_value'}
+    description = 'custom description'
+    series = init_series(sample_series,
+                         logical_type='categorical',
+                         semantic_tags=['custom_tag'],
+                         metadata=metadata,
+                         description=description,
+                         use_standard_tags=False)
+    assert series is not sample_series
+    correct_dtype = Categorical._get_valid_dtype(type(sample_series))
+    assert series.dtype == correct_dtype
+    assert isinstance(series.ww.logical_type, Categorical)
+    assert series.ww.semantic_tags == {'custom_tag'}
+    assert series.ww.metadata == metadata
+    assert series.ww.description == description
